+++ test/fixtures/routeProps.js
@@ -1,0 +1,28 @@
+const routeProps = {
+  history: {
+    action: 'PUSH',
+    block: jest.fn(),
+    createHref: jest.fn(),
+    go: jest.fn(),
+    listen: jest.fn(),
+    location: {
+      hash: '',
+      pathname: '',
+      search: '',
+    },
+    push: () => jest.fn(),
+    replace: () => jest.fn(),
+  },
+  location: {
+    hash: '',
+    pathname: '',
+    search: '',
+  },
+  match: {
+    params: {
+      id: '9a2427cd-4110-4bd9-b6f9-e3475631bbac',
+    },
+  },
+};
+
+export default routeProps;